Move getBlockDescription and getSnippetName into new file
